Fix updating path, break filling text from button
Tried to github copilot my way out of the path property not saving when text is updated, and it kinda works but introduces a couple bugs (like the textbox blanking itself after you edit, changes not saving until you go to the next row...)
But it also breaks the button's ability to fill out the path now.

I noticed it works in BorwseField.xaml if you change this line:
                 Text="{Binding Path, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
to this:
                 Text="{Binding Text, RelativeSource={RelativeSource AncestorType=UserControl}, UpdateSourceTrigger=PropertyChanged}"
Which updates the UI when you use the buttonto select a file again, but you lose the ability to save changes to the path to the object. Fun.